Fix first task on new server being very slow
Essentially, gradle is downloading all of its dependencies when the first
job on a new server is run. Instead, we make it download everything when
the image is built, so that it is there already when real jobs start
running. This is achieved by making it compile a hello world program.